fix undocumented id attribute crashing parsing video json
Starting on the morning of the 27th of september, the `videos`
attribute in the reponse to requests on a media endpoint contain
a new attribute: id, which always set to '0'. This crashed the
`Video.object_from_dictionary` calls for every response parsed by the
client that contained this tag as the id attribute wasn't present on
the Video model.

e.g.

```javascript

"videos": {
    "standard_resolution": {
	"width": 640,
	    "height": 359,
	    "url": "https://scontent.cdninstagram.com/t50.2886-16/21347497_688135741377726_1513723350453583872_n.mp4",
	    "id": "0" // wat.
    },
	"low_bandwidth": {
	    "width": 480,
	    "height": 269,
	    "url": "https://scontent.cdninstagram.com/t50.2886-16/21364131_115181835834691_686212731677704192_n.mp4",
	    "id": "0"
	},
	"low_resolution": {
	    "width": 480,
	    "height": 269,
	    "url": "https://scontent.cdninstagram.com/t50.2886-16/21364131_115181835834691_686212731677704192_n.mp4",
	    "id": "0"
	}
}

```

As of now, we have no idea what this attribute is supposed to signify
(it's not documented, nor is anything mentioned in the changelog) and
chances are this is a mistake on instagram's end. We've made it so the
Video model should be able to deal with the response changing back
should instagram remove the attribute again.

We've accounted for the two scenarios (the tag being there and it not
being there) in the tests by having one of the two video representations
have it and the other one not.